workaround setup selinux https://github.com/ansible/ansible/issues/16612
``ansible_selinux`` is not always consistent; it may be either a dict with a
``status`` entry, or a boolean set to false.

This is fixed in ansible 2.4.1 (see link).

This commit enforces a first check on ansible_selinux before checking
ansible_selinux.enabled. It enables role to work with pre and post
ansible 2.4.1 as python non-empty dict is true for boolean checking.